requirements.txt
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.25.3
pandas >= 1.1.4